apply renames to code
